refactor/fix discord modules
