+++ src/interfaces/discord/discord.module.ts
@@ -1,23 +1,21 @@
-import { Logger, Module } from '@nestjs/common';
+import { Module } from '@nestjs/common';
 import { DiscordPingService } from './services/discord-ping.service';
 import { NecordModule } from 'necord';
 
 import { DiscordConfigService } from './services/discord-config.service';
 import { DiscordEventService } from './services/events/discord-event.service';
-import { DiscordChatService } from './services/chat-command/discord-chat.service';
 import { GcpModule } from '../../core/gcp/gcp.module';
 import { UsersModule } from '../../core/database/collections/users/users.module';
 import { ChatbotModule } from '../../core/chatbot/chatbot.module';
 import { AiChatsModule } from '../../core/database/collections/ai-chats/ai-chats.module';
 import { CoupletsModule } from '../../core/database/collections/couplets/couplets.module';
 import { MessagesModule } from '../../core/database/collections/messages/messages.module';
-import { DiscordContextService } from './services/chat-command/discord-context.service';
+import { DiscordContextService } from './services/discord-context.service';
 import { DiscordMongodbService } from './services/discord-mongodb.service';
-import { DiscordMessageService } from './services/chat-command/discord-message.service';
+import { DiscordMessageService } from './services/discord-message.service';
 import { OpenaiModule } from '../../core/ai/openai/openai.module';
-import { DiscordAttachmentService } from './services/chat-command/discord-attachment.service';
+import { DiscordAttachmentService } from './services/discord-attachment.service';
 import { DiscordOnMessageService } from './services/events/discord-on-message.service';
-import { DiscordCommandsModule } from './commands/discord-commands.module';
 
 @Module({
   imports: [
@@ -32,18 +30,16 @@ import { DiscordCommandsModule } from './commands/discord-commands.module';
     CoupletsModule,
     MessagesModule,
     OpenaiModule,
-    DiscordCommandsModule,
   ],
   providers: [
-    DiscordOnMessageService,
     DiscordEventService,
+    DiscordOnMessageService,
     DiscordPingService,
-    DiscordChatService,
     DiscordMessageService,
     DiscordAttachmentService,
     DiscordContextService,
     DiscordMongodbService,
-    Logger,
   ],
+  exports: [DiscordMongodbService],
 })
 export class DiscordModule {}
