--- src/interfaces/discord/dto/discord-text.dto.ts
@@ -1,10 +1,0 @@
-import { StringOption } from 'necord';
-
-export class DiscordTextDto {
-  @StringOption({
-    name: 'text',
-    description: 'Your text',
-    required: false,
-  })
-  text: string;
-}
